app/requirements.txt
pandas
numpy
flask
PyYAML
requests
json
logging
jwt